Add template creation and use a json decoder for input for bulks apis (#5751)
Template creation will be via a lazy initialization in order to save
calls to Elastic when ingesting data. A template will be linked with the
ILM policy, mappings, index pattern and create a boostrap index.

* Add debug logs

* Run gofumpt

* Some logging tweak and bug fixes

* [CODE REVIEW] Cosmetic changes and add support for base settings

* Adjust logging

* Update application name to fluentd

* [Code review] Name changes for templates and mappings

* Add tenant

+++ linseed/pkg/apis/v1/l3_flow.go
@@ -167,7 +167,7 @@ type Process struct {
 type Service struct {
 	Name      string `json:"name"`
 	Namespace string `json:"namespace"`
-	Port      int32  `json:"port"`
+	Port      int64  `json:"port"`
 	PortName  string `json:"port_name"`
 }
